MAINT: create private modules when they should (#452)
closes #422
