+++ imblearn/combine/__init__.py
@@ -2,7 +2,7 @@
 over-sampling and under-sampling.
 """
 
-from .smote_enn import SMOTEENN
-from .smote_tomek import SMOTETomek
+from ._smote_enn import SMOTEENN
+from ._smote_tomek import SMOTETomek
 
 __all__ = ['SMOTEENN', 'SMOTETomek']
