+++ imblearn/combine/_smote_enn.py
@@ -1,5 +1,4 @@
-"""Class to perform over-sampling using SMOTE and cleaning using Tomek
-links."""
+"""Class to perform over-sampling using SMOTE and cleaning using ENN."""
 
 # Authors: Guillaume Lemaitre <g.lemaitre58@gmail.com>
 #          Christos Aridas
@@ -8,7 +7,6 @@ links."""
 from __future__ import division
 
 import logging
-import warnings
 
 from sklearn.base import clone
 from sklearn.utils import check_X_y
@@ -16,7 +14,7 @@ from sklearn.utils import check_X_y
 from ..base import BaseSampler
 from ..over_sampling import SMOTE
 from ..over_sampling.base import BaseOverSampler
-from ..under_sampling import TomekLinks
+from ..under_sampling import EditedNearestNeighbours
 from ..utils import check_target_type, hash_X_y
 from ..utils import Substitution
 from ..utils._docstring import _random_state_docstring
@@ -25,11 +23,10 @@ from ..utils._docstring import _random_state_docstring
 @Substitution(
     sampling_strategy=BaseOverSampler._sampling_strategy_docstring,
     random_state=_random_state_docstring)
-class SMOTETomek(BaseSampler):
-    """Class to perform over-sampling using SMOTE and cleaning using
-    Tomek links.
+class SMOTEENN(BaseSampler):
+    """Class to perform over-sampling using SMOTE and cleaning using ENN.
 
-    Combine over- and under-sampling using SMOTE and Tomek links.
+    Combine over- and under-sampling using SMOTE and Edited Nearest Neighbours.
 
     Read more in the :ref:`User Guide <combine>`.
 
@@ -44,11 +41,6 @@ class SMOTETomek(BaseSampler):
         a :class:`imblearn.over_sampling.SMOTE` object with default parameters
         will be given.
 
-    tomek : object, optional (default=Tomek())
-        The :class:`imblearn.under_sampling.Tomek` object to use. If not given,
-        a :class:`imblearn.under_sampling.Tomek` object with default parameters
-        will be given.
-
     ratio : str, dict, or callable
         .. deprecated:: 0.4
            Use the parameter ``sampling_strategy`` instead. It will be removed
@@ -56,40 +48,40 @@ class SMOTETomek(BaseSampler):
 
     Notes
     -----
-    The methos is presented in [1]_.
+    The method is presented in [1]_.
 
-    Supports multi-class resampling. Refer to SMOTE and TomekLinks regarding
-    the scheme which used.
+    Supports multi-class resampling. Refer to SMOTE and ENN regarding the
+    scheme which used.
 
-    See :ref:`sphx_glr_auto_examples_combine_plot_smote_tomek.py` and
+    See :ref:`sphx_glr_auto_examples_combine_plot_smote_enn.py` and
     :ref:`sphx_glr_auto_examples_combine_plot_comparison_combine.py`.
 
     See also
     --------
-    SMOTEENN : Over-sample using SMOTE followed by under-sampling using Edited
-        Nearest Neighbours.
+    SMOTETomek : Over-sample using SMOTE followed by under-sampling removing
+        the Tomek's links.
 
     References
     ----------
-    .. [1] G. Batista, B. Bazzan, M. Monard, "Balancing Training Data for
-       Automated Annotation of Keywords: a Case Study," In WOB, 10-18, 2003.
+    .. [1] G. Batista, R. C. Prati, M. C. Monard. "A study of the behavior of
+       several methods for balancing machine learning training data," ACM
+       Sigkdd Explorations Newsletter 6 (1), 20-29, 2004.
 
     Examples
     --------
 
     >>> from collections import Counter
     >>> from sklearn.datasets import make_classification
-    >>> from imblearn.combine import \
-SMOTETomek # doctest: +NORMALIZE_WHITESPACE
+    >>> from imblearn.combine import SMOTEENN # doctest: +NORMALIZE_WHITESPACE
     >>> X, y = make_classification(n_classes=2, class_sep=2,
     ... weights=[0.1, 0.9], n_informative=3, n_redundant=1, flip_y=0,
     ... n_features=20, n_clusters_per_class=1, n_samples=1000, random_state=10)
     >>> print('Original dataset shape %s' % Counter(y))
     Original dataset shape Counter({{1: 900, 0: 100}})
-    >>> smt = SMOTETomek(random_state=42)
-    >>> X_res, y_res = smt.fit_sample(X, y)
+    >>> sme = SMOTEENN(random_state=42)
+    >>> X_res, y_res = sme.fit_sample(X, y)
     >>> print('Resampled dataset shape %s' % Counter(y_res))
-    Resampled dataset shape Counter({{0: 900, 1: 900}})
+    Resampled dataset shape Counter({{0: 900, 1: 881}})
 
     """
 
@@ -97,19 +89,18 @@ SMOTETomek # doctest: +NORMALIZE_WHITESPACE
                  sampling_strategy='auto',
                  random_state=None,
                  smote=None,
-                 tomek=None,
+                 enn=None,
                  ratio=None):
-        super(SMOTETomek, self).__init__()
+        super(SMOTEENN, self).__init__()
         self.sampling_strategy = sampling_strategy
         self.random_state = random_state
         self.smote = smote
-        self.tomek = tomek
+        self.enn = enn
         self.ratio = ratio
         self.logger = logging.getLogger(__name__)
 
     def _validate_estimator(self):
         "Private function to validate SMOTE and ENN objects"
-
         if self.smote is not None:
             if isinstance(self.smote, SMOTE):
                 self.smote_ = clone(self.smote)
@@ -123,15 +114,15 @@ SMOTETomek # doctest: +NORMALIZE_WHITESPACE
                 random_state=self.random_state,
                 ratio=self.ratio)
 
-        if self.tomek is not None:
-            if isinstance(self.tomek, TomekLinks):
-                self.tomek_ = clone(self.tomek)
+        if self.enn is not None:
+            if isinstance(self.enn, EditedNearestNeighbours):
+                self.enn_ = clone(self.enn)
             else:
-                raise ValueError('tomek needs to be a TomekLinks object.'
-                                 'Got {} instead.'.format(type(self.tomek)))
-        # Otherwise create a default TomekLinks
+                raise ValueError('enn needs to be an EditedNearestNeighbours.'
+                                 ' Got {} instead.'.format(type(self.enn)))
+        # Otherwise create a default EditedNearestNeighbours
         else:
-            self.tomek_ = TomekLinks(sampling_strategy='all')
+            self.enn_ = EditedNearestNeighbours(sampling_strategy='all')
 
     def fit(self, X, y):
         """Find the classes statistics before to perform sampling.
@@ -174,11 +165,11 @@ SMOTETomek # doctest: +NORMALIZE_WHITESPACE
 (n_samples_new, n_features)
             The array containing the resampled data.
 
-        y_resampled : ndarray, shape (n_samples_new,)
+        y_resampled : ndarray, shape (n_samples_new)
             The corresponding label of `X_resampled`
 
         """
         self._validate_estimator()
 
         X_res, y_res = self.smote_.fit_sample(X, y)
-        return self.tomek_.fit_sample(X_res, y_res)
+        return self.enn_.fit_sample(X_res, y_res)
