+++ imblearn/combine/_smote_tomek.py
@@ -1,4 +1,5 @@
-"""Class to perform over-sampling using SMOTE and cleaning using ENN."""
+"""Class to perform over-sampling using SMOTE and cleaning using Tomek
+links."""
 
 # Authors: Guillaume Lemaitre <g.lemaitre58@gmail.com>
 #          Christos Aridas
@@ -7,7 +8,6 @@
 from __future__ import division
 
 import logging
-import warnings
 
 from sklearn.base import clone
 from sklearn.utils import check_X_y
@@ -15,7 +15,7 @@ from sklearn.utils import check_X_y
 from ..base import BaseSampler
 from ..over_sampling import SMOTE
 from ..over_sampling.base import BaseOverSampler
-from ..under_sampling import EditedNearestNeighbours
+from ..under_sampling import TomekLinks
 from ..utils import check_target_type, hash_X_y
 from ..utils import Substitution
 from ..utils._docstring import _random_state_docstring
@@ -24,10 +24,11 @@ from ..utils._docstring import _random_state_docstring
 @Substitution(
     sampling_strategy=BaseOverSampler._sampling_strategy_docstring,
     random_state=_random_state_docstring)
-class SMOTEENN(BaseSampler):
-    """Class to perform over-sampling using SMOTE and cleaning using ENN.
+class SMOTETomek(BaseSampler):
+    """Class to perform over-sampling using SMOTE and cleaning using
+    Tomek links.
 
-    Combine over- and under-sampling using SMOTE and Edited Nearest Neighbours.
+    Combine over- and under-sampling using SMOTE and Tomek links.
 
     Read more in the :ref:`User Guide <combine>`.
 
@@ -42,6 +43,11 @@ class SMOTEENN(BaseSampler):
         a :class:`imblearn.over_sampling.SMOTE` object with default parameters
         will be given.
 
+    tomek : object, optional (default=Tomek())
+        The :class:`imblearn.under_sampling.Tomek` object to use. If not given,
+        a :class:`imblearn.under_sampling.Tomek` object with default parameters
+        will be given.
+
     ratio : str, dict, or callable
         .. deprecated:: 0.4
            Use the parameter ``sampling_strategy`` instead. It will be removed
@@ -49,40 +55,40 @@ class SMOTEENN(BaseSampler):
 
     Notes
     -----
-    The method is presented in [1]_.
+    The methos is presented in [1]_.
 
-    Supports multi-class resampling. Refer to SMOTE and ENN regarding the
-    scheme which used.
+    Supports multi-class resampling. Refer to SMOTE and TomekLinks regarding
+    the scheme which used.
 
-    See :ref:`sphx_glr_auto_examples_combine_plot_smote_enn.py` and
+    See :ref:`sphx_glr_auto_examples_combine_plot_smote_tomek.py` and
     :ref:`sphx_glr_auto_examples_combine_plot_comparison_combine.py`.
 
     See also
     --------
-    SMOTETomek : Over-sample using SMOTE followed by under-sampling removing
-        the Tomek's links.
+    SMOTEENN : Over-sample using SMOTE followed by under-sampling using Edited
+        Nearest Neighbours.
 
     References
     ----------
-    .. [1] G. Batista, R. C. Prati, M. C. Monard. "A study of the behavior of
-       several methods for balancing machine learning training data," ACM
-       Sigkdd Explorations Newsletter 6 (1), 20-29, 2004.
+    .. [1] G. Batista, B. Bazzan, M. Monard, "Balancing Training Data for
+       Automated Annotation of Keywords: a Case Study," In WOB, 10-18, 2003.
 
     Examples
     --------
 
     >>> from collections import Counter
     >>> from sklearn.datasets import make_classification
-    >>> from imblearn.combine import SMOTEENN # doctest: +NORMALIZE_WHITESPACE
+    >>> from imblearn.combine import \
+SMOTETomek # doctest: +NORMALIZE_WHITESPACE
     >>> X, y = make_classification(n_classes=2, class_sep=2,
     ... weights=[0.1, 0.9], n_informative=3, n_redundant=1, flip_y=0,
     ... n_features=20, n_clusters_per_class=1, n_samples=1000, random_state=10)
     >>> print('Original dataset shape %s' % Counter(y))
     Original dataset shape Counter({{1: 900, 0: 100}})
-    >>> sme = SMOTEENN(random_state=42)
-    >>> X_res, y_res = sme.fit_sample(X, y)
+    >>> smt = SMOTETomek(random_state=42)
+    >>> X_res, y_res = smt.fit_sample(X, y)
     >>> print('Resampled dataset shape %s' % Counter(y_res))
-    Resampled dataset shape Counter({{0: 900, 1: 881}})
+    Resampled dataset shape Counter({{0: 900, 1: 900}})
 
     """
 
@@ -90,18 +96,19 @@ class SMOTEENN(BaseSampler):
                  sampling_strategy='auto',
                  random_state=None,
                  smote=None,
-                 enn=None,
+                 tomek=None,
                  ratio=None):
-        super(SMOTEENN, self).__init__()
+        super(SMOTETomek, self).__init__()
         self.sampling_strategy = sampling_strategy
         self.random_state = random_state
         self.smote = smote
-        self.enn = enn
+        self.tomek = tomek
         self.ratio = ratio
         self.logger = logging.getLogger(__name__)
 
     def _validate_estimator(self):
         "Private function to validate SMOTE and ENN objects"
+
         if self.smote is not None:
             if isinstance(self.smote, SMOTE):
                 self.smote_ = clone(self.smote)
@@ -115,15 +122,15 @@ class SMOTEENN(BaseSampler):
                 random_state=self.random_state,
                 ratio=self.ratio)
 
-        if self.enn is not None:
-            if isinstance(self.enn, EditedNearestNeighbours):
-                self.enn_ = clone(self.enn)
+        if self.tomek is not None:
+            if isinstance(self.tomek, TomekLinks):
+                self.tomek_ = clone(self.tomek)
             else:
-                raise ValueError('enn needs to be an EditedNearestNeighbours.'
-                                 ' Got {} instead.'.format(type(self.enn)))
-        # Otherwise create a default EditedNearestNeighbours
+                raise ValueError('tomek needs to be a TomekLinks object.'
+                                 'Got {} instead.'.format(type(self.tomek)))
+        # Otherwise create a default TomekLinks
         else:
-            self.enn_ = EditedNearestNeighbours(sampling_strategy='all')
+            self.tomek_ = TomekLinks(sampling_strategy='all')
 
     def fit(self, X, y):
         """Find the classes statistics before to perform sampling.
@@ -166,11 +173,11 @@ class SMOTEENN(BaseSampler):
 (n_samples_new, n_features)
             The array containing the resampled data.
 
-        y_resampled : ndarray, shape (n_samples_new)
+        y_resampled : ndarray, shape (n_samples_new,)
             The corresponding label of `X_resampled`
 
         """
         self._validate_estimator()
 
         X_res, y_res = self.smote_.fit_sample(X, y)
-        return self.enn_.fit_sample(X_res, y_res)
+        return self.tomek_.fit_sample(X_res, y_res)
