+++ imblearn/datasets/__init__.py
@@ -3,8 +3,8 @@ The :mod:`imblearn.datasets` provides methods to generate
 imbalanced data.
 """
 
-from .imbalance import make_imbalance
+from ._imbalance import make_imbalance
 
-from .zenodo import fetch_datasets
+from ._zenodo import fetch_datasets
 
 __all__ = ['make_imbalance', 'fetch_datasets']
