+++ imblearn/under_sampling/__init__.py
@@ -3,18 +3,18 @@ The :mod:`imblearn.under_sampling` provides methods to under-sample
 a dataset.
 """
 
-from .prototype_generation import ClusterCentroids
+from ._prototype_generation import ClusterCentroids
 
-from .prototype_selection import RandomUnderSampler
-from .prototype_selection import TomekLinks
-from .prototype_selection import NearMiss
-from .prototype_selection import CondensedNearestNeighbour
-from .prototype_selection import OneSidedSelection
-from .prototype_selection import NeighbourhoodCleaningRule
-from .prototype_selection import EditedNearestNeighbours
-from .prototype_selection import RepeatedEditedNearestNeighbours
-from .prototype_selection import AllKNN
-from .prototype_selection import InstanceHardnessThreshold
+from ._prototype_selection import RandomUnderSampler
+from ._prototype_selection import TomekLinks
+from ._prototype_selection import NearMiss
+from ._prototype_selection import CondensedNearestNeighbour
+from ._prototype_selection import OneSidedSelection
+from ._prototype_selection import NeighbourhoodCleaningRule
+from ._prototype_selection import EditedNearestNeighbours
+from ._prototype_selection import RepeatedEditedNearestNeighbours
+from ._prototype_selection import AllKNN
+from ._prototype_selection import InstanceHardnessThreshold
 
 __all__ = [
     'ClusterCentroids', 'RandomUnderSampler', 'InstanceHardnessThreshold',
