+++ imblearn/under_sampling/_prototype_generation/__init__.py
@@ -3,6 +3,6 @@ The :mod:`imblearn.under_sampling.prototype_generation` submodule contains
 methods that generate new samples in order to balance the dataset.
 """
 
-from .cluster_centroids import ClusterCentroids
+from ._cluster_centroids import ClusterCentroids
 
 __all__ = ['ClusterCentroids']
