+++ imblearn/under_sampling/_prototype_selection/__init__.py
@@ -1,0 +1,22 @@
+"""
+The :mod:`imblearn.under_sampling.prototype_selection` submodule contains
+methods that select samples in order to balance the dataset.
+"""
+
+from ._random_under_sampler import RandomUnderSampler
+from ._tomek_links import TomekLinks
+from ._nearmiss import NearMiss
+from ._condensed_nearest_neighbour import CondensedNearestNeighbour
+from ._one_sided_selection import OneSidedSelection
+from ._neighbourhood_cleaning_rule import NeighbourhoodCleaningRule
+from ._edited_nearest_neighbours import EditedNearestNeighbours
+from ._edited_nearest_neighbours import RepeatedEditedNearestNeighbours
+from ._edited_nearest_neighbours import AllKNN
+from ._instance_hardness_threshold import InstanceHardnessThreshold
+
+__all__ = [
+    'RandomUnderSampler', 'InstanceHardnessThreshold', 'NearMiss',
+    'TomekLinks', 'EditedNearestNeighbours', 'RepeatedEditedNearestNeighbours',
+    'AllKNN', 'OneSidedSelection', 'CondensedNearestNeighbour',
+    'NeighbourhoodCleaningRule'
+]
