--- imblearn/under_sampling/prototype_selection/__init__.py
@@ -1,22 +1,0 @@
-"""
-The :mod:`imblearn.under_sampling.prototype_selection` submodule contains
-methods that select samples in order to balance the dataset.
-"""
-
-from .random_under_sampler import RandomUnderSampler
-from .tomek_links import TomekLinks
-from .nearmiss import NearMiss
-from .condensed_nearest_neighbour import CondensedNearestNeighbour
-from .one_sided_selection import OneSidedSelection
-from .neighbourhood_cleaning_rule import NeighbourhoodCleaningRule
-from .edited_nearest_neighbours import EditedNearestNeighbours
-from .edited_nearest_neighbours import RepeatedEditedNearestNeighbours
-from .edited_nearest_neighbours import AllKNN
-from .instance_hardness_threshold import InstanceHardnessThreshold
-
-__all__ = [
-    'RandomUnderSampler', 'InstanceHardnessThreshold', 'NearMiss',
-    'TomekLinks', 'EditedNearestNeighbours', 'RepeatedEditedNearestNeighbours',
-    'AllKNN', 'OneSidedSelection', 'CondensedNearestNeighbour',
-    'NeighbourhoodCleaningRule'
-]
